[css] 이미지 생성 페이지 완성
